requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.10
typing-extensions>=4.9.0 